Change page-indexing size_t to uint32_t
Some actual consistancy. Page limit seems to be always handled with 32-bit integers in Sai anyways.

+++ libsai/sai.hpp
@@ -104,10 +104,10 @@ public:
 private:
 	std::ifstream FileIn;
 
-	std::size_t PageCount;
+	std::uint32_t PageCount;
 
 	VirtualPage Buffer;
-	std::size_t CurrentPage;
+	std::uint32_t CurrentPage;
 
 	// Decryption Key
 	const std::uint32_t *Key;
